moved JSON serialization source into its own extension file - inspired by the std. library the JSON serialize and deserialize source is located under json/add/...
